test(core): migrate TodoShortcutCatalogTests (Wave D)
4 tests, 22 failure sites preserved 1:1. The shortcutSection helper's `throw TestFailure` converts to `try #require`.
